Platform has been modified to compute its bounding box when the model changes rather than relying on the dead reackoning helper to do so. This solves an issue where flaming entities are initialized high in the air when the application starts up with those entities with flames enabled.
From analyzing the code with Brad, we found that the particle system that is attached to the entity affects the initial bounding box computation, thus affecting the placement of the 3 isectors used by the entity for ground clamping; the particle system seems to have its world offset while the entity's position is placed back at the origin for the initial bounding box calculation in the dead reckoning helper, thus resulting in an extremely large bounding box. With a large bounding box, the 3 isectors are placed far apart which obtain very different values and their Z values are averaged for the final Z position for the entity gournd clamping.