Nulled out the user data on NxActors that are being deleted.  That really didn't help the crash and reload issue.
Now the physics component deletes and recreates all the scenes on map unload and on restarted.   It remembers all group collision flags for the user, but other stuff must be added as a callback that it will call when after the scenes have all be recreated.  It saves the original scene desc so that it can use that to rebuild the scene, and it will store the a few values such as flags and gravity so they can be reset too.