Fixed some problems with the generic config xml file.
Fixed the stealth viewer hud init so things startup in the right places and don't need a resize to move them.

The label manager, after the last commit, was calculating the font size every frame, now it just does it once for each label, or when the text changes, as before.

Fixed the human to have the collision volume in the right place.

Fixed the platform with physics to use kinematic physics when remote.

Fixed some collision groups:  bullets shouldn't really collide with anything, vehicle should collide with local humans, but not remote ones.