Added new Dead Reckoning config options: "SimCore.DR.UseCubicSpline" and "SimCore.DR.UseFixedTimeBlends". The first one allows you to force DR to use the old linear velocity blend, instead of splines. The second one allows you to force DR to calc an avg update time for blending. If not specified, these options do not auto-default to anything, letting Delta do whatever it did before. Loaded in OnEnteredWorld in BaseEntity.
BasePhysicsVehicle now sets the DR algorithm to Velocity only for brand new actors - this setting is overridden by map or network properties. Also, moved the setting of UseMaxSmoothing Time to allow the above config option to work.

DRPublishingActComp - Fixed a bug - ForceUpdateNextChance was allowing you to bypass the MaxUpdateSendRate value. Fixed a bug - setting the MaxUpdateSendRate forced the actor to set the AlwaysUseMaxSmoothingTime on the DRHelper, which means most entities ended up with a 0.2sec blend time. Fixed the order of operations for the calc when using VelocityAveragingFrameCount > 1. 

Deleted some cruft commented code in Platform.